Name some test files more meaningfully
